Lighting Bolt
-Added lighting bolt spell, needs centering
-Adjusted lighting for level 1 and 2
-Added lighting for slimes for visibility
-Hasty Added a tutorial prompt at level 1
