+++ Assets/Animation/Lighting.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1eb13e46a34553d4f818d6514c54008f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
