+++ Assets/Animation/Lighting/LightningBolt.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: eb4e9c9d1d279c049bc3f7b61913631d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
