+++ Assets/Animation/Lighting/Strike.anim
@@ -1,0 +1,105 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!74 &7400000
+AnimationClip:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: Strike
+  serializedVersion: 7
+  m_Legacy: 0
+  m_Compressed: 0
+  m_UseHighQualityCurve: 1
+  m_RotationCurves: []
+  m_CompressedRotationCurves: []
+  m_EulerCurves: []
+  m_PositionCurves: []
+  m_ScaleCurves: []
+  m_FloatCurves: []
+  m_PPtrCurves:
+  - serializedVersion: 2
+    curve:
+    - time: 0
+      value: {fileID: -536906485, guid: 3ff480d9ba617dc40a32f2d5cb3de59c, type: 3}
+    - time: 0.05
+      value: {fileID: 516493966, guid: 3ff480d9ba617dc40a32f2d5cb3de59c, type: 3}
+    - time: 0.06666667
+      value: {fileID: 294732332, guid: 3ff480d9ba617dc40a32f2d5cb3de59c, type: 3}
+    - time: 0.06666667
+      value: {fileID: -1416254997, guid: 3ff480d9ba617dc40a32f2d5cb3de59c, type: 3}
+    - time: 0.11666667
+      value: {fileID: -1975311844, guid: 3ff480d9ba617dc40a32f2d5cb3de59c, type: 3}
+    - time: 0.16666667
+      value: {fileID: 754673356, guid: 3ff480d9ba617dc40a32f2d5cb3de59c, type: 3}
+    - time: 0.18333334
+      value: {fileID: 1917362840, guid: 3ff480d9ba617dc40a32f2d5cb3de59c, type: 3}
+    - time: 0.21666667
+      value: {fileID: 1370985086, guid: 3ff480d9ba617dc40a32f2d5cb3de59c, type: 3}
+    - time: 0.23333333
+      value: {fileID: 2109626, guid: 3ff480d9ba617dc40a32f2d5cb3de59c, type: 3}
+    - time: 0.23333333
+      value: {fileID: -2136300145, guid: 3ff480d9ba617dc40a32f2d5cb3de59c, type: 3}
+    - time: 0.26666668
+      value: {fileID: 947148960, guid: 3ff480d9ba617dc40a32f2d5cb3de59c, type: 3}
+    - time: 0.3
+      value: {fileID: 508431630, guid: 3ff480d9ba617dc40a32f2d5cb3de59c, type: 3}
+    attribute: m_Sprite
+    path: 
+    classID: 212
+    script: {fileID: 0}
+    flags: 2
+  m_SampleRate: 60
+  m_WrapMode: 0
+  m_Bounds:
+    m_Center: {x: 0, y: 0, z: 0}
+    m_Extent: {x: 0, y: 0, z: 0}
+  m_ClipBindingConstant:
+    genericBindings:
+    - serializedVersion: 2
+      path: 0
+      attribute: 0
+      script: {fileID: 0}
+      typeID: 212
+      customType: 23
+      isPPtrCurve: 1
+      isIntCurve: 0
+      isSerializeReferenceCurve: 0
+    pptrCurveMapping:
+    - {fileID: -536906485, guid: 3ff480d9ba617dc40a32f2d5cb3de59c, type: 3}
+    - {fileID: 516493966, guid: 3ff480d9ba617dc40a32f2d5cb3de59c, type: 3}
+    - {fileID: 294732332, guid: 3ff480d9ba617dc40a32f2d5cb3de59c, type: 3}
+    - {fileID: -1416254997, guid: 3ff480d9ba617dc40a32f2d5cb3de59c, type: 3}
+    - {fileID: -1975311844, guid: 3ff480d9ba617dc40a32f2d5cb3de59c, type: 3}
+    - {fileID: 754673356, guid: 3ff480d9ba617dc40a32f2d5cb3de59c, type: 3}
+    - {fileID: 1917362840, guid: 3ff480d9ba617dc40a32f2d5cb3de59c, type: 3}
+    - {fileID: 1370985086, guid: 3ff480d9ba617dc40a32f2d5cb3de59c, type: 3}
+    - {fileID: 2109626, guid: 3ff480d9ba617dc40a32f2d5cb3de59c, type: 3}
+    - {fileID: -2136300145, guid: 3ff480d9ba617dc40a32f2d5cb3de59c, type: 3}
+    - {fileID: 947148960, guid: 3ff480d9ba617dc40a32f2d5cb3de59c, type: 3}
+    - {fileID: 508431630, guid: 3ff480d9ba617dc40a32f2d5cb3de59c, type: 3}
+  m_AnimationClipSettings:
+    serializedVersion: 2
+    m_AdditiveReferencePoseClip: {fileID: 0}
+    m_AdditiveReferencePoseTime: 0
+    m_StartTime: 0
+    m_StopTime: 0.3166667
+    m_OrientationOffsetY: 0
+    m_Level: 0
+    m_CycleOffset: 0
+    m_HasAdditiveReferencePose: 0
+    m_LoopTime: 0
+    m_LoopBlend: 0
+    m_LoopBlendOrientation: 0
+    m_LoopBlendPositionY: 0
+    m_LoopBlendPositionXZ: 0
+    m_KeepOriginalOrientation: 0
+    m_KeepOriginalPositionY: 1
+    m_KeepOriginalPositionXZ: 0
+    m_HeightFromFeet: 0
+    m_Mirror: 0
+  m_EditorCurves: []
+  m_EulerEditorCurves: []
+  m_HasGenericRootTransform: 0
+  m_HasMotionFloatCurves: 0
+  m_Events: []
