+++ Assets/Animation/Lighting/Strike.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: eb0d7dc89f2f4a74fa93014593b06c2b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
