+++ Assets/Art/Effects/Lighting.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a04fbe9479decbd42a0895a8dc6d5761
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
