+++ Assets/Objects/Entities/LightningBolt.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 23a1f824b15c6b149a9b4b60c7aef3b9
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
