+++ Assets/ProximityTextHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f2e365230dd512b45ac9783eb2038171
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
