+++ Assets/Scripts/C# Scripts/Environment/PlayerDevour.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d3eb97f288b20f94db12cace5a84079b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
