+++ Assets/Scripts/C# Scripts/Lighting.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d4ac0cca174243f4f927810191dd670b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
