+++ Assets/Scripts/C# Scripts/Lighting/LightingBolt.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 234806430d835d84fa4cd9c6e5f049c6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
