--- Assets/TeleportingSlimeEnemy.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c1fa1b20963960f499f06fc8f53661de
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
